join fn refactor
